requirements.txt
certifi==2024.12.14
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
yt-dlp==2024.12.23
